Moved DataSizeFormatter to the common module.
