+++ common/src/test/java/com/duckduckgo/app/global/formatters/time/DatabaseDateFormatterTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 DuckDuckGo
+ * Copyright (c) 2022 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.mobile.android.vpn.store
+package com.duckduckgo.app.global.formatters.time
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
